easy coding -
ThreadLocal分而治之：
1 用法, 特性 与 原理；
2 使用 InheritableThreadLocal 实现 不同线程之间的共享变量透传；
3 线程池中使用 threadLocal可能导致的脏数据问题 & 解决方案remove()
