+++ src/main/java/com/henry/tryout/easy_coding/concurrency_and_multiple_thread/ThreadLocal_05/threadlocal_usage_02/InheritableThreadLocal_and_SimpleDateFormat_02/RequestProcessTrace_03.java
@@ -1,7 +1,8 @@
-package com.henry.tryout.easy_coding.concurrency_and_multiple_thread.ThreadLocal_05.threadlocal_usage_02;
+package com.henry.tryout.easy_coding.concurrency_and_multiple_thread.ThreadLocal_05.threadlocal_usage_02.InheritableThreadLocal_and_SimpleDateFormat_02;
 
 import org.apache.commons.lang.StringUtils;
 
+// 在父子线程之间 共享线程变量 - 手段：InheritableThreadLocal
 public class RequestProcessTrace_03 {
     // 定义了一个 InheritableThreadLocal类型的变量 - 封装了 FullLinkContext类型变量
     private static final InheritableThreadLocal<FullLinkContext> FULL_LINK_THREADLOCAL
@@ -12,7 +13,7 @@ public class RequestProcessTrace_03 {
         // get对象
         FullLinkContext fullLinkContext = FULL_LINK_THREADLOCAL.get();
         if (fullLinkContext != null) {
-            // set回对象
+            // set新的对象
             FULL_LINK_THREADLOCAL.set(new FullLinkContext());
             // 把set后的对象 绑定回 原始引用
             fullLinkContext = FULL_LINK_THREADLOCAL.get();
