+++ src/main/java/com/henry/tryout/easy_coding/concurrency_and_multiple_thread/ThreadLocal_05/threadlocal_usage_02/InheritableThreadLocal_and_SimpleDateFormat_02/SimpleDateFormatDemo.java
@@ -1,4 +1,4 @@
-package com.henry.tryout.easy_coding.concurrency_and_multiple_thread.ThreadLocal_05.threadlocal_usage_02;
+package com.henry.tryout.easy_coding.concurrency_and_multiple_thread.ThreadLocal_05.threadlocal_usage_02.InheritableThreadLocal_and_SimpleDateFormat_02;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
